Stability update
Really small change, sometimes readStringUntil() will return null with the example code and cause the processing application to hang. Adding a simple if statement will drop the frame and alert the user in the console of the issue so the demo can continue to run. Tested working with ESP32 + GridEYE + Latest Processing 3.5.4.